+++ Editor/Playa/SaintsFieldWithInfo.cs
@@ -15,6 +15,7 @@ namespace SaintsField.Editor.Playa
         public IReadOnlyList<IPlayaAttribute> PlayaAttributes;
         public object TargetParent;
         public MemberInfo TargetMemberInfo;
+        public int TargetMemberIndex;
         public IReadOnlyList<object> Targets;
 
         public SaintsRenderType RenderType;