+++ Samples~/Scripts/IssueAndTesting/Issue/Issue240SaintsRowStructChange/FooComponent.cs
@@ -1,6 +1,5 @@
 using System;
 using SaintsField.Playa;
-using SaintsField.Samples.Scripts.SaintsEditor;
 
 namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue.Issue240SaintsRowStructChange
 {
@@ -17,10 +16,6 @@ namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue.Issue240SaintsRowStr
             private void FooInline()
             {
                 fooInline++;
-// #if UNITY_EDITOR
-//                 SaintsContext.SerializedProperty.intValue++;  // for inline buttons we get the decorated property
-//                 SaintsContext.SerializedProperty.serializedObject.ApplyModifiedProperties();
-// #endif
             }
 
             public int foo;
@@ -29,11 +24,6 @@ namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue.Issue240SaintsRowStr
             private void IncrementFoo()
             {
                 foo++;
-// #if UNITY_EDITOR
-//                 // for `Button` we get the property of the container (`foo1` field here)
-//                 SaintsContext.FindPropertyRelateTo(nameof(foo)).intValue++;
-//                 SaintsContext.SerializedProperty.serializedObject.ApplyModifiedProperties();
-// #endif
             }
 
             public override string ToString()
@@ -41,30 +31,5 @@ namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue.Issue240SaintsRowStr
                 return $"<Foo1 {fooInline}, {foo}>";
             }
         }
-
-//         [Serializable]
-//         public struct FloatWithBaseValue
-//         {
-//             [OnValueChanged(nameof(RevertFinalToBaseValue))] public float baseValue;
-//             public float finalValue;
-//
-//             public void RevertFinalToBaseValue()
-//             {
-//                 finalValue = baseValue;
-//             }
-//
-//             [OnValueChanged(nameof(RevertFinalToBaseValue2))] public float baseValue2;
-//             public float finalValue2;
-//
-//             public void RevertFinalToBaseValue2()
-//             {
-// #if UNITY_EDITOR
-//                 SaintsContext.FindPropertyRelateTo(nameof(finalValue2)).floatValue = baseValue2;
-//                 SaintsContext.SerializedProperty.serializedObject.ApplyModifiedProperties();
-// #endif
-//             }
-//         }
-//
-//         public FloatWithBaseValue fwb;
     }
 }